+++ src/frequenz/client/common/__init__.py
@@ -5,12 +5,16 @@
 
 from ._exception import (
     ClientCommonError,
+    InvalidAttributeError,
+    MissingFieldError,
     UnrecognizedEnumValueError,
     UnspecifiedEnumValueError,
 )
 
 __all__ = [
     "ClientCommonError",
+    "InvalidAttributeError",
+    "MissingFieldError",
     "UnrecognizedEnumValueError",
     "UnspecifiedEnumValueError",
 ]